salesflow-ai/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
openai==1.52.2
anthropic>=0.18.0
supabase==2.6.0
pytest==8.3.3
httpx>=0.25.0
